Massive breakthrough: Cast API JSON response to int instead of float, and return strictly int 0 or 1

+++ server/app.py
@@ -66,7 +66,7 @@ def step(req: StepRequest):
 
     return {
         "question": result.observation.question,
-        "reward": float(result.reward),
+        "reward": int(result.reward),
         "done": result.done
     }
